streamline data_manager with how the databases are used
