+++ utils/data/__init__.py
@@ -1,5 +1,5 @@
 # from ._database import client, db_mogis, db_players, db_mogis
-from .data_manager import data_manager, archive_type
+from .data_manager import data_manager
 from .mogi_manager import mogi_manager
 from .roombrowser import get_room_info, ServerType
 from .state import state_manager
@@ -12,7 +12,6 @@ from .table_reader_api import (
 
 __all__ = [
     "data_manager",
-    "archive_type",
     "mogi_manager",
     "get_room_info",
     "ServerType",
